module and session
added check for session expiration and now sending available modules to the client

+++ API/Loader/functions/auth/session.php
@@ -3,7 +3,7 @@ namespace auth;
 
 function generate_session($connection, $aes_key) {
     $session_id= \rnd_string_secure(32);
-    $session_expiration = time() + 300;
+    $session_expiration = time() + 300; // 5 minutes 
 
     $connection->query('INSERT INTO loader_sessions(session_id,enc_key,expiration) VALUES(?,?)',[$session_id,$aes_key,$session_expiration]);
 
@@ -12,13 +12,26 @@ function generate_session($connection, $aes_key) {
            ];
 }
 
+/* 
+
+0 - expired session
+
+*/
+
 function get_session_from_id($connection, $session_id,$fetch_all) {
     $query = $connection->query('SELECT * FROM loader_sessions WHERE session_id=? LIMIT 1',[$session_id]);
 
-    if ($fetch_all)
-        return $query->fetch_assoc();
+    $row_data = $query->fetch_assoc();
 
-    return $query->fetch_assoc()['enc_key'];
+    if (time() > $row_data['expiration']) {
+        destroy_session($connection,$session_id);
+        return 0;
+    }
+
+    if ($fetch_all)
+        return $row_data;
+        
+    return $row_data['enc_key'];
 }
 
 function add_session_db_identifiers($connection, $session_id,$username,$loader_key) {
@@ -26,8 +39,11 @@ function add_session_db_identifiers($connection, $session_id,$username,$loader_k
     $connection->query('UPDATE loader_sessions SET loader_key=? WHERE session_id=?',[$loader_key,$session_id]);
 }
 
-function destroy_session($connection, $session_id,$loader_key) {
-    $connection->query('DELETE FROM loader_sessions WHERE session_id=? AND loader_key=?',[$session_id,$loader_key]);
+function destroy_session($connection, $session_id,$loader_key = "",$owner = "") {
+    if ($loader_key !== "" && $owner !== "")
+        $connection->query('DELETE FROM loader_sessions WHERE session_id=? AND loader_key=? AND owner=?',[$session_id,$loader_key,$owner]);
+        
+    $connection->query('DELETE FROM loader_sessions WHERE session_id=?'[$session_id]);
 }
 
 ?>
